Got the dotnet app working with the K6 scripts
