--- dotnet-performance-app/dotnet-performance-app/dotnet-performance-app.csproj
@@ -1,36 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <RootNamespace>dotnet_performance_app</RootNamespace>
-    <UserSecretsId>b8c268d7-49e0-456a-8646-5b80f062d25c</UserSecretsId>
-    <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
-    <DockerfileContext>.</DockerfileContext>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="fo-dicom" Version="5.0.2" />
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.14.0" />
-    <PackageReference Include="Monai.Deploy.Messaging" Version="0.1.0-rc0040" />
-    <PackageReference Include="RabbitMQ.Client" Version="6.2.4" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Data\DICOM\CT\" />
-    <Folder Include="Data\DICOM\ULTRASOUND\" />
-    <Folder Include="Data\DICOM\XRAY\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="Data\DICOM\MRI\Patient001\MR000000.dcm">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Update="Data\DICOM\MRI\Patient002\MR000000.dcm">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
