+++ performance-testing/docker-compose.yml
@@ -1,0 +1,18 @@
+version: "3.2"
+services:
+  rabbitmq:
+    image: rabbitmq:3-management-alpine
+    container_name: 'rabbitmq'
+    environment:
+      RABBITMQ_DEFAULT_USER: "admin"
+      RABBITMQ_DEFAULT_PASS: "admin"
+      RABBITMQ_DEFAULT_VHOST: "monaideploy"
+    ports:
+        - 5672:5672
+        - 15672:15672
+  orthanc:
+    image: jodogne/orthanc:latest
+    container_name: 'orthanc'
+    ports:
+        - 4242:4242
+        - 8042:8042
