+++ performance-testing/dotnet-performance-app/.dockerignore
