+++ performance-testing/dotnet-performance-app/Data/DICOM/CT/Patient001/MR000000.dcm
