+++ performance-testing/dotnet-performance-app/Data/DICOM/CT/Patient002/MR000000.dcm
