+++ performance-testing/dotnet-performance-app/Support/DicomScu.cs
