+++ performance-testing/dotnet-performance-app/Support/RabbitPublisher.cs
