--- performance-testing/k6-with-rabbit-and-dicom/docker-compose.yml
@@ -1,14 +1,0 @@
-version: "3.2"
-services:
-  rabbitmq:
-    image: rabbitmq:3-management-alpine
-    container_name: 'rabbitmq'
-    ports:
-        - 5672:5672
-        - 15672:15672
-  orthanc:
-    image: jodogne/orthanc:latest
-    container_name: 'orthanc'
-    ports:
-        - 4242:4242
-        - 8042:8042
