--- performance-testing/k6-with-rabbit-and-dicom/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "rabbitmq-with-k6",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "start": "nodemon ./bin/www",
-    "test": "k6 run tests/k6.js",
-    "test:automatic": "start-server-and-test start http://localhost:3000 test "
-  },
-  "dependencies": {
-    "@cloudamqp/amqp-client": "^2.0.3",
-    "cookie-parser": "~1.4.4",
-    "debug": "~2.6.9",
-    "dicom-dimse-native": "^2.3.2",
-    "express": "~4.16.1",
-    "http-errors": "~1.6.3",
-    "jade": "~1.11.0",
-    "morgan": "~1.9.1",
-    "nodemon": "^2.0.15",
-    "start-server-and-test": "^1.14.0"
-  },
-  "devDependencies": {
-    "@types/node": "^17.0.31"
-  }
-}
