--- performance-testing/k6-with-rabbit-and-dicom/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "compilerOptions": {
-        "outDir": "./dist",
-        "allowJs": true,
-        "target": "es5"
-    },
-        "include": [
-        "./*"
-    ]
-}
