--- performance-testing/k6-with-rabbit-and-dicom/views/index.jade
@@ -1,5 +1,0 @@
-extends layout
-
-block content
-  h1= title
-  p Welcome to #{title}
