+++ performance-testing/k6/dicom/config/avgConfig.json
