+++ performance-testing/k6/dicom/config/peakConfig.json
